+++ tests/integration/mod.rs
@@ -1,0 +1,3 @@
+mod iroh;
+#[cfg(feature = "tcp")]
+mod tcp;
