+++ tests/piglet/tcp.rs
@@ -1,0 +1,11 @@
+use super::super::{ip_port, kill, run, wait_for_output};
+use serial_test::serial;
+
+#[test]
+#[serial]
+fn ip_is_output() {
+    let mut child = run("piglet", vec![], None);
+    let line = wait_for_output(&mut child, "ip:").expect("Could not get ip");
+    kill(child);
+    let (_, _) = ip_port(&line);
+}
